Split the Template tests
