+++ eng/pipelines/handlers.yml
@@ -88,7 +88,7 @@ parameters:
         vmImage: $(macOSXVmImage)
         artifact: nuget-macos
 
-  - name: BuilTemplatePlatforms
+  - name: BuildTemplatePlatforms
     type: object
     default:
       - name: Windows
@@ -273,7 +273,7 @@ stages:
                 PRIVATE_BUILD: $(PrivateBuild)
             - pwsh: ./build.ps1 --target=dotnet-integration-build --verbosity=diagnostic
               displayName: Build Microsoft.Maui.IntegrationTests
-            - pwsh: ./build.ps1 --target=dotnet-integration-test --filter="FullyQualifiedName=Microsoft.Maui.IntegrationTests.SampleTests" --resultsfilename="integration-samples" --verbosity=diagnostic
+            - pwsh: ./build.ps1 --target=dotnet-integration-test --filter="Category=Samples" --resultsfilename="integration-samples" --verbosity=diagnostic
               displayName: Run ${{ BuildPlatform.name }} sample build tests
             - task: PublishTestResults@2
               displayName: Publish the ${{ BuildPlatform.name }} sample build tests
@@ -295,7 +295,7 @@ stages:
     - template: common/maui-templates.yml
       parameters:
         RunPlatforms: ${{ parameters.RunTemplatePlatforms }}
-        BuildPlatforms: ${{ parameters.BuilTemplatePlatforms }}
+        BuildPlatforms: ${{ parameters.BuildTemplatePlatforms }}
 
   # - ${{ if eq(variables['System.TeamProject'], 'devdiv') }}:
   #   - template: common/localization-handoff.yml                     # Process outgoing strings [Localization Handoff]
